refactor: separate reader state from component behavior
This refactor decouples the reader's persistent state from its interactive behavior. The Alpine store is now dedicated to state management (theme, font, fullscreen), while a new `readerControls` component handles UI-specific logic such as scroll tracking and toolbar visibility. This improves code organization and follows Alpine.js best practices for separating state from behavior.

Changes:

- Modified `resources/ts/app.ts`: Added import for global stores.
- Modified `resources/ts/components/story/index.ts`: Updated reader component entry point.
- Created `resources/ts/components/story/reader-controls.ts`: Implemented Alpine component for reader UI logic and scroll handling.
- Deleted `resources/ts/components/story/reader.ts`: Removed old combined store and logic implementation.
- Created `resources/ts/stores/index.ts`: Added entry point for Alpine stores.
- Created `resources/ts/stores/reader.ts`: Extracted reader state into a dedicated Alpine store.
- Modified `resources/views/components/story/reader-controls.blade.php`: Updated to use `readerControls` data and event-based scroll logic.
- Modified `resources/views/livewire/story/view-story.blade.php`: Updated scroll listeners to dispatch `reader-scroll` events.
