+++ resources/ts/app.ts
@@ -1,3 +1,4 @@
 import "./components";
 import "./events";
 import "./navigation-menu";
+import "./stores";
