+++ resources/ts/components/story/index.ts
@@ -1,1 +1,1 @@
-import "./reader";
+import "./reader-controls";
